Change maths for milli-amp/hour counting

+++ ATTINYCellModule/src/packet_processor.cpp
@@ -335,7 +335,7 @@ bool PacketProcessor::processPacket(PacketStruct *buffer)
 
   case COMMAND::ReadBalanceCurrentCounter:
   {
-    buffer->moduledata[mymoduleaddress] = MilliAmpHourBalanceCounter;
+    buffer->moduledata[mymoduleaddress] = (uint16_t)MilliAmpHourBalanceCounter;
     return true;
   }
